+++ backends/simplego/buffers.go
@@ -213,22 +213,26 @@ func copyFlat(flatDst, flatSrc any) {
 }
 
 // mutableBytes returns the slice of the bytes used by the flat given -- it works with any of the supported data types for buffers.
-func (b *Buffer) mutableBytes() []byte {
-	fn := mutableBytesDTypeMap.Get(b.shape.DType).(func(b *Buffer) []byte) //nolint:errcheck
+func (b *Buffer) mutableBytes() ([]byte, error) {
+	tmpAny, tmpErr := mutableBytesDTypeMap.Get(b.shape.DType)
+	if tmpErr != nil {
+		return nil, tmpErr
+	}
+	fn := tmpAny.(func(b *Buffer) ([]byte, error)) //nolint:errcheck
 	return fn(b)
 }
 
 var mutableBytesDTypeMap = NewDTypeMap("MutableBytes")
 
 // mutableBytesGeneric is the generic implementation of mutableBytes.
-func mutableBytesGeneric[T SupportedTypesConstraints](b *Buffer) []byte {
+func mutableBytesGeneric[T SupportedTypesConstraints](b *Buffer) ([]byte, error) {
 	flat := b.flat.([]T) //nolint:errcheck
 	if len(flat) == 0 {
-		return nil // Handle empty tensors
+		return nil, nil // Handle empty tensors
 	}
 	bytePointer := (*byte)(unsafe.Pointer(&flat[0]))
 	var t T
-	return unsafe.Slice(bytePointer, len(flat)*int(unsafe.Sizeof(t)))
+	return unsafe.Slice(bytePointer, len(flat)*int(unsafe.Sizeof(t))), nil
 }
 
 // Fill the buffer with the given value.
@@ -240,7 +244,11 @@ func (b *Buffer) Fill(value any) error {
 	if value != nil && dtypes.FromAny(value) != dtype {
 		return errors.Errorf("fillBuffer: invalid value type %T for buffer of dtype %s", value, dtype)
 	}
-	fillFn := fillBufferDTypeMap.Get(dtype).(func(*Buffer, any)) //nolint:errcheck
+	tmpAny, tmpErr := fillBufferDTypeMap.Get(dtype)
+	if tmpErr != nil {
+		panic(tmpErr)
+	}
+	fillFn := tmpAny.(func(*Buffer, any)) //nolint:errcheck
 	fillFn(b, value)
 	return nil
 }